+++ example/lib/main.dart
@@ -50,9 +50,34 @@ class MyApp extends SPA {
 }
 
 class Counter extends Component {
+  late Animation<double> _animation;
+  late AnimationController _controller;
+
   int _count = 0;
 
-  void increment() => setState(() => _count++);
+  Future<void> increment() async {
+    await _controller.reverse();
+    _count++;
+    await _controller.forward();
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    _controller = AnimationController(
+        duration: const Duration(milliseconds: 200),
+        reverseDuration: const Duration(milliseconds: 200),
+        vsync: vsync);
+    _animation = Tween<double>(begin: 0, end: 1).animate(_controller)
+      ..addListener(() => setState(() {}));
+    _controller.forward();
+  }
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
 
   @override
   Widget render(BuildContext context) => Column(
@@ -64,9 +89,14 @@ class Counter extends Component {
           Text(
             '$_count',
             style: Theme.of(context).textTheme.headlineMedium,
+            textScaleFactor: _animation.value,
           ),
           ElevatedButton(
-            onPressed: () => setState(() => _count = 0),
+            onPressed: () {
+              _count = 0;
+              _controller.reset();
+              _controller.fling();
+            },
             child: const Text('Reset'),
           ),
         ],